+++ src/compounds/billboard/CHANGELOG.md
@@ -3,6 +3,47 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.4](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.billboard@0.1.3...@uswitch/trustyle.billboard@0.1.4) (2021-04-29)
+
+
+### Bug Fixes
+
+* restyled width and color for primary slot ([1752c96](https://github.com/uswitch/trustyle/commit/1752c96))
+
+
+
+
+
+## [0.1.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.billboard@0.1.2...@uswitch/trustyle.billboard@0.1.3) (2021-04-29)
+
+
+### Bug Fixes
+
+* changed spacing and font color ([d169be2](https://github.com/uswitch/trustyle/commit/d169be2))
+
+
+
+
+
+## [0.1.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.billboard@0.1.1...@uswitch/trustyle.billboard@0.1.2) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.billboard
+
+
+
+
+
+## [0.1.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.billboard@0.1.0...@uswitch/trustyle.billboard@0.1.1) (2021-04-26)
+
+
+### Bug Fixes
+
+* adjusted spacing for breadcrumb and primary content ([5f29a17](https://github.com/uswitch/trustyle/commit/5f29a17))
+
+
+
+
+
 # [0.1.0](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.billboard@0.0.1...@uswitch/trustyle.billboard@0.1.0) (2021-04-23)