+++ src/elements/bullet-list-highlight/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.16](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.bullet-list-highlight@0.4.15...@uswitch/trustyle.bullet-list-highlight@0.4.16) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.bullet-list-highlight
+
+
+
+
+
 ## [0.4.15](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.bullet-list-highlight@0.4.14...@uswitch/trustyle.bullet-list-highlight@0.4.15) (2021-04-21)
 
 **Note:** Version bump only for package @uswitch/trustyle.bullet-list-highlight