+++ src/elements/call-out/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.22](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.call-out@2.1.21...@uswitch/trustyle.call-out@2.1.22) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.call-out
+
+
+
+
+
 ## [2.1.21](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.call-out@2.1.20...@uswitch/trustyle.call-out@2.1.21) (2021-04-21)
 
 **Note:** Version bump only for package @uswitch/trustyle.call-out