+++ src/elements/card/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.12.19](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.12.18...@uswitch/trustyle.card@1.12.19) (2021-04-29)
+
+**Note:** Version bump only for package @uswitch/trustyle.card
+
+
+
+
+
+## [1.12.18](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.12.17...@uswitch/trustyle.card@1.12.18) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.card
+
+
+
+
+
 ## [1.12.17](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.card@1.12.16...@uswitch/trustyle.card@1.12.17) (2021-04-23)
 
 **Note:** Version bump only for package @uswitch/trustyle.card