+++ src/elements/drop-down/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.5.10](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.drop-down@2.5.9...@uswitch/trustyle.drop-down@2.5.10) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.drop-down
+
+
+
+
+
 ## [2.5.9](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.drop-down@2.5.8...@uswitch/trustyle.drop-down@2.5.9) (2021-04-21)
 
 **Note:** Version bump only for package @uswitch/trustyle.drop-down