+++ src/elements/progress-circle/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.progress-circle@1.1.2...@uswitch/trustyle.progress-circle@1.1.3) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.progress-circle
+
+
+
+
+
 # 1.1.0 (2020-09-14)