+++ src/elements/read-more-card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.18](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.read-more-card@1.4.17...@uswitch/trustyle.read-more-card@1.4.18) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.read-more-card
+
+
+
+
+
 ## [1.4.17](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.read-more-card@1.4.16...@uswitch/trustyle.read-more-card@1.4.17) (2021-04-21)
 
 **Note:** Version bump only for package @uswitch/trustyle.read-more-card