+++ src/elements/simple-icon/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.1](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.simple-icon@0.1.0...@uswitch/trustyle.simple-icon@0.1.1) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.simple-icon
+
+
+
+
+
 # 0.1.0 (2021-02-09)