+++ src/elements/simple-input/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.3](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.simple-input@0.6.2...@uswitch/trustyle.simple-input@0.6.3) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.simple-input
+
+
+
+
+
 ## [0.6.2](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.simple-input@0.6.1...@uswitch/trustyle.simple-input@0.6.2) (2021-02-25)
 
 **Note:** Version bump only for package @uswitch/trustyle.simple-input