+++ src/themes/bankrate/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.25.13](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.bankrate-theme@2.25.12...@uswitch/trustyle.bankrate-theme@2.25.13) (2021-04-27)
+
+**Note:** Version bump only for package @uswitch/trustyle.bankrate-theme
+
+
+
+
+
 ## [2.25.12](https://github.com/uswitch/trustyle/compare/@uswitch/trustyle.bankrate-theme@2.25.11...@uswitch/trustyle.bankrate-theme@2.25.12) (2021-03-09)